Move lazy evaluation to separate file
* Remove context from interpreter Thunk
